Docs {tm}	R Documentation

Access Document IDs and Terms

Description

Accessing document IDs, terms, and their number of a term-document matrix or document-term matrix.

Usage

Docs(x)
nDocs(x)
nTerms(x)
Terms(x)

Arguments

x
Either a TermDocumentMatrix or DocumentTermMatrix.

Value

For Docs and Terms, a character vector with document IDs and terms, respectively.

For nDocs and nTerms, an integer with the number of document IDs and terms, respectively.

Examples

data("crude")
tdm <- TermDocumentMatrix(crude)[1:10,1:20]
Docs(tdm)
nDocs(tdm)
nTerms(tdm)
Terms(tdm)

[Package tm version 0.7-5 Index]