VectorSource {tm}	R Documentation

Vector Source

Description

Create a vector source.

Usage

VectorSource(x)

Arguments

x
A vector giving the texts.

Details

A vector source interprets each element of the vector x as a document.

Value

An object inheriting from VectorSource, SimpleSource, and Source.

See Also

Source for basic information on the source infrastructure employed by package tm.

Examples

docs <- c("This is a text.", "This another one.")
(vs <- VectorSource(docs))
inspect(VCorpus(vs))

[Package tm version 0.7-5 Index]