WeightFunction {tm}	R Documentation

Weighting Function

Description

Construct a weighting function for term-document matrices.

Usage

WeightFunction(x, name, acronym)

Arguments

x
A function which takes a TermDocumentMatrix with term frequencies as input, weights the elements, and returns the weighted matrix.

name
A character naming the weighting function.

acronym
A character giving an acronym for the name of the weighting function.

Value

An object of class WeightFunction which extends the class function representing a weighting function.

Examples

weightCutBin <- WeightFunction(function(m, cutoff) m > cutoff,
                               "binary with cutoff", "bincut")

[Package tm version 0.7-5 Index]