XMLSource {tm}	R Documentation

XML Source

Description

Create an XML source.

Usage

XMLSource(x, parser = xml_contents, reader)

Arguments

x
a character giving a uniform resource identifier.

parser
a function accepting an XML document (as delivered by read_xml in package xml2) as input and returning XML elements/nodes.

reader
a function capable of turning XML elements/nodes as returned by parser into a subclass of TextDocument.

Value

An object inheriting from XMLSource, SimpleSource, and Source.

See Also

Source for basic information on the source infrastructure employed by package tm.

Vignette 'Extensions: How to Handle Custom File Formats', and readXML.

[Package tm version 0.7-5 Index]