XMLTextDocument {tm}	R Documentation

XML Text Documents

Description

Create XML text documents.

Usage

XMLTextDocument(x = xml_missing(),
                author = character(0),
                datetimestamp = as.POSIXlt(Sys.time(), tz = "GMT"),
                description = character(0),
                heading = character(0),
                id = character(0),
                language = character(0),
                origin = character(0),
                ...,
                meta = NULL)

Arguments

x
An XMLDocument.

author
a character or an object of class person giving the author names.

datetimestamp
an object of class POSIXt or a character string giving the creation date/time information. If a character string, exactly one of the ISO 8601 formats defined by http://www.w3.org/TR/NOTE-datetime should be used. See parse_ISO_8601_datetime in package NLP for processing such date/time information.

description
a character giving a description.

heading
a character giving the title or a short heading.

id
a character giving a unique identifier.

language
a character giving the language (preferably as IETF language tags, see language in package NLP).

origin
a character giving information on the source and origin.

...
user-defined document metadata tag-value pairs.

meta
a named list or NULL (default) giving all metadata. If set all other metadata arguments are ignored.

Value

An object inheriting from XMLTextDocument and TextDocument.

See Also

TextDocument for basic information on the text document infrastructure employed by package tm.

Examples

xml <- system.file("extdata", "order-doc.xml", package = "xml2")
(xtd <- XMLTextDocument(xml2::read_xml(xml),
                        heading = "XML text document",
                        id = xml,
                        language = "en"))
content(xtd)
meta(xtd)

[Package tm version 0.7-5 Index]