ZipSource {tm}	R Documentation

ZIP File Source

Description

Create a ZIP file source.

Usage

ZipSource(zipfile,
          pattern = NULL,
          recursive = FALSE,
          ignore.case = FALSE,
          mode = "text")

Arguments

zipfile
A character string with the full path name of a ZIP file.

pattern
an optional regular expression. Only file names in the ZIP file which match the regular expression will be returned.

recursive
logical. Should the listing recurse into directories?

ignore.case
logical. Should pattern-matching be case-insensitive?

mode
a character string specifying if and how files should be read in. Available modes are:

""

No read. In this case getElem and pGetElem only deliver URIs.

"binary"

Files are read in binary raw mode (via readBin).

"text"

Files are read as text (via readLines).

Details

A ZIP file source extracts a compressed ZIP file via unzip and interprets each file as a document.

Value

An object inheriting from ZipSource, SimpleSource, and Source.

See Also

Source for basic information on the source infrastructure employed by package tm.

Examples

zipfile <- tempfile()
files <- Sys.glob(file.path(system.file("texts", "txt", package = "tm"), "*"))
zip(zipfile, files)
zipfile <- paste0(zipfile, ".zip")
Corpus(ZipSource(zipfile, recursive = TRUE))[[1]]
file.remove(zipfile)

[Package tm version 0.7-5 Index]