acq {tm}	R Documentation

50 Exemplary News Articles from the Reuters-21578 Data Set of Topic acq

Description

This dataset holds 50 news articles with additional meta information from the Reuters-21578 data set. All documents belong to the topic acq dealing with corporate acquisitions.

Usage

data("acq")

Format

A VCorpus of 50 text documents.

Source

Reuters-21578 Text Categorization Collection Distribution 1.0 (XML format).

References

Emms, Martin and Luz, Saturnino (2007). Machine Learning for Natural Language Processing. European Summer School of Logic, Language and Information, course reader. https://www.scss.tcd.ie/~luzs/publications/mlfornlp.pdf

Lewis, David (1997) Reuters-21578 Text Categorization Collection Distribution 1.0. http://kdd.ics.uci.edu/databases/reuters21578/reuters21578.html

Luz, Saturnino XML-encoded version of Reuters-21578. https://www.scss.tcd.ie/~luzs/t/cs4ll4/sw/reuters21578-xml/

Examples

data("acq")
acq

[Package tm version 0.7-5 Index]