findFreqTerms {tm}	R Documentation

Find Frequent Terms

Description

Find frequent terms in a document-term or term-document matrix.

Usage

findFreqTerms(x, lowfreq = 0, highfreq = Inf)

Arguments

x
A DocumentTermMatrix or TermDocumentMatrix.

lowfreq
A numeric for the lower frequency bound.

highfreq
A numeric for the upper frequency bound.

Details

This method works for all numeric weightings but is probably most meaningful for the standard term frequency (tf) weighting of x.

Value

A character vector of terms in x which occur more or equal often than lowfreq times and less or equal often than highfreq times.

Examples

data("crude")
tdm <- TermDocumentMatrix(crude)
findFreqTerms(tdm, 2, 3)

[Package tm version 0.7-5 Index]