plot {tm}	R Documentation

Visualize a Term-Document Matrix

Description

Visualize correlations between terms of a term-document matrix.

Usage

## S3 method for class 'TermDocumentMatrix'
plot(x,
     terms = sample(Terms(x), 20),
     corThreshold = 0.7,
     weighting = FALSE,
     attrs = list(graph = list(rankdir = "BT"),
                  node = list(shape = "rectangle",
                              fixedsize = FALSE)),
     ...)

Arguments

x
A term-document matrix.

terms
Terms to be plotted. Defaults to 20 randomly chosen terms of the term-document matrix.

corThreshold
Do not plot correlations below this threshold. Defaults to 0.7.

weighting
Define whether the line width corresponds to the correlation.

attrs
Argument passed to the plot method for class graphNEL.

...
Other arguments passed to the graphNEL plot method.

Details

Visualization requires that package Rgraphviz is available.

Examples

## Not run: data(crude)
tdm <- TermDocumentMatrix(crude,
                          control = list(removePunctuation = TRUE,
                                         removeNumbers = TRUE,
                                         stopwords = TRUE))
plot(tdm, corThreshold = 0.2, weighting = TRUE)
## End(Not run)

[Package tm version 0.7-5 Index]