readDataframe {tm}	R Documentation

Read In a Text Document from a Data Frame

Description

Read in a text document from a row in a data frame.

Usage

readDataframe(elem, language, id)

Arguments

elem
a named list with the component content which must hold a data frame with rows as the documents to be read in. The names of the columns holding the text content and the document identifier must be "text" and "doc_id", respectively.

language
a string giving the language.

id
Not used.

Value

A PlainTextDocument representing elem$content.

See Also

Reader for basic information on the reader infrastructure employed by package tm.

Examples

docs <- data.frame(doc_id = c("doc_1", "doc_2"),
                   text = c("This is a text.", "This another one."),
                   stringsAsFactors = FALSE)
ds <- DataframeSource(docs)
elem <- getElem(stepNext(ds))
result <- readDataframe(elem, "en", NULL)
inspect(result)
meta(result)

[Package tm version 0.7-5 Index]