weightTf {tm}	R Documentation

Weight by Term Frequency

Description

Weight a term-document matrix by term frequency.

Usage

weightTf(m)

Arguments

m
A TermDocumentMatrix in term frequency format.

Details

Formally this function is of class WeightingFunction with the additional attributes name and acronym.

This function acts as the identity function since the input matrix is already in term frequency format.

Value

The weighted matrix.

[Package tm version 0.7-5 Index]